+++ packages/bridge-controller/src/types.ts
@@ -1,4 +1,4 @@
-import type { AccountsControllerGetSelectedMultichainAccountAction } from '@metamask/accounts-controller';
+import type { AccountsControllerGetAccountByAddressAction } from '@metamask/accounts-controller';
 import type {
   GetCurrencyRateState,
   MultichainAssetsRatesControllerGetStateAction,
@@ -30,6 +30,7 @@ import type { BRIDGE_CONTROLLER_NAME } from './constants/bridge';
 import type {
   BridgeAssetSchema,
   ChainConfigurationSchema,
+  FeatureId,
   FeeDataSchema,
   PlatformConfigSchema,
   ProtocolSchema,
@@ -39,19 +40,9 @@ import type {
   TxDataSchema,
 } from './utils/validators';
 
-/**
- * Additional options accepted by the extension's fetchWithCache function
- */
-type FetchWithCacheOptions = {
-  cacheOptions?: {
-    cacheRefreshTime: number;
-  };
-  functionName?: string;
-};
-
 export type FetchFunction = (
-  input: RequestInfo | URL,
-  init?: RequestInit & FetchWithCacheOptions,
+  input: RequestInfo | URL | string,
+  init?: RequestInit,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
 ) => Promise<any>;
 
@@ -78,8 +69,8 @@ export type L1GasFees = {
   l1GasFeesInHexWei?: string; // l1 fees for approval and trade in hex wei, appended by BridgeController.#appendL1GasFees
 };
 
-export type SolanaFees = {
-  solanaFeesInLamports?: string; // solana fees in lamports, appended by BridgeController.#appendSolanaFees
+export type NonEvmFees = {
+  nonEvmFeesInNative?: string; // Non-EVM chain fees in native units (SOL for Solana, BTC for Bitcoin)
 };
 
 /**
@@ -265,6 +256,7 @@ export type QuoteResponse<TxDataType = TxData> = Infer<
 > & {
   trade: TxDataType;
   approval?: TxData;
+  featureId?: FeatureId;
 };
 
 export enum ChainId {
@@ -303,11 +295,31 @@ export enum BridgeBackgroundAction {
 
 export type BridgeControllerState = {
   quoteRequest: Partial<GenericQuoteRequest>;
-  quotes: (QuoteResponse & L1GasFees & SolanaFees)[];
+  quotes: (QuoteResponse & L1GasFees & NonEvmFees)[];
+  /**
+   * The time elapsed between the initial quote fetch and when the first valid quote was received
+   */
   quotesInitialLoadTime: number | null;
+  /**
+   * The timestamp of when the latest quote fetch started
+   */
   quotesLastFetched: number | null;
+  /**
+   * The status of the quote fetch, including fee calculations and validations
+   * This is set to
+   * - LOADING when the quote fetch starts
+   * - FETCHED when the process completes successfully, including when quotes are empty
+   * - ERROR when any errors occur
+   *
+   * When SSE is enabled, this is set to LOADING even when a quote is available. It is only
+   * set to FETCHED when the stream is closed and all quotes have been received
+   */
   quotesLoadingStatus: RequestStatus | null;
   quoteFetchError: string | null;
+  /**
+   * The number of times the quotes have been refreshed, starts at 0 and is
+   * incremented at the end of each quote fetch
+   */
   quotesRefreshCount: number;
   /**
    * Asset exchange rates for EVM and multichain assets that are not indexed by the assets controllers
@@ -351,7 +363,7 @@ export type BridgeControllerActions =
 export type BridgeControllerEvents = BridgeControllerStateChangeEvent;
 
 export type AllowedActions =
-  | AccountsControllerGetSelectedMultichainAccountAction
+  | AccountsControllerGetAccountByAddressAction
   | GetCurrencyRateState
   | TokenRatesControllerGetStateAction
   | MultichainAssetsRatesControllerGetStateAction